refactor: move http services to its own folder, move types to separate files where appropriate
