+++ src/components/user/controller/createUserAction.ts
@@ -1,13 +1,13 @@
 import { constants as httpConstants } from 'node:http2';
-import { Request } from '../../../framework/Request';
-import { Response } from '../../../framework/Response';
-import { HandlerFn } from '../../../framework/Router';
+import { Request } from '../../../framework/http/Request';
+import { Response } from '../../../framework/http/Response';
+import { ActionHandler } from '../../../framework/http/Router';
 import { userRepository } from '../userRepository';
 import { validateModel } from '../../../framework/validator';
 import { User } from '../userEntity';
 import { CreateUserDto, validationRules } from '../userDto';
 
-const createUser: HandlerFn = async (request: Request, response: Response): Promise<void> => {
+const createUser: ActionHandler = async (request: Request, response: Response): Promise<void> => {
     const newUserDto = request.getJsonBody();
 
     validateModel<CreateUserDto>(newUserDto, validationRules);
