+++ src/components/user/controller/getUserAction.ts
@@ -1,12 +1,12 @@
 import { constants as httpConstants } from 'node:http2';
-import { HandlerFn } from '../../../framework/Router';
-import { Request } from '../../../framework/Request';
-import { Response } from '../../../framework/Response';
+import { ActionHandler } from '../../../framework/http/Router';
+import { Request } from '../../../framework/http/Request';
+import { Response } from '../../../framework/http/Response';
 import { userRepository } from '../userRepository';
 import { assertValidUuid } from '../../../framework/asserts';
 import { USER_NOT_FOUND } from '../exceptionMessages';
 
-const getUser: HandlerFn = async (request: Request, response: Response): Promise<void> => {
+const getUser: ActionHandler = async (request: Request, response: Response): Promise<void> => {
     const { id } = request.getPlaceholderValues();
 
     assertValidUuid(id);
