+++ src/components/user/controller/getUsersAction.ts
@@ -1,9 +1,9 @@
-import { Request } from '../../../framework/Request';
-import { Response } from '../../../framework/Response';
-import { HandlerFn } from '../../../framework/Router';
+import { Request } from '../../../framework/http/Request';
+import { Response } from '../../../framework/http/Response';
+import { ActionHandler } from '../../../framework/http/Router';
 import { userRepository } from '../userRepository';
 
-const getUsers: HandlerFn = async (_: Request, response: Response): Promise<void> => {
+const getUsers: ActionHandler = async (_: Request, response: Response): Promise<void> => {
     const users = await userRepository.findAll();
 
     response.json(users);
