+++ src/framework/http/ActionHandler.ts
@@ -1,0 +1,6 @@
+import { Request } from './Request';
+import { Response } from './Response';
+
+type ActionHandler = (request: Request, response: Response) => void;
+
+export { ActionHandler };
