+++ src/framework/http/HttpMethodEnum.ts
@@ -1,0 +1,8 @@
+enum HttpMethod {
+    Get = 'GET',
+    Post = 'POST',
+    Put = 'PUT',
+    Delete = 'DELETE',
+}
+
+export { HttpMethod };
